Move form to core, file-input to root, fix regex, CSS, clean dom graphical elements
