+++ src/window.ts
@@ -1,4 +1,4 @@
 export const getSearchWord = (): string | undefined => {
-  const res = /\?(\w+)/g.exec(window.location.search);
+  const res = /\?([\w-]+)/g.exec(window.location.search);
   return res ? res[1] : undefined;
 }
